refactor: 훅 이동
